댓글 기능과 css부분들 수정 그리고 기본적인 validation부븐을 추가

+++ src/main/java/cf/myboard/www/mapper/BoardMapper.java
@@ -10,7 +10,7 @@ public interface BoardMapper {
 
     List<Board> getBoardList(Board board);
 
-    int countBoard(Board board);
+    long countBoard(Board board);
 
     Board getBoard(long id);
